+++ src/Utils/FilepondUtil.php
@@ -31,25 +31,47 @@ class FilepondUtil
     }
 
     /**
+     * Convert the request to prioritize file first
+     *
+     * @see https://github.com/laravel/framework/issues/61356
+     *
+     * @return array
+     */
+    public static function convertRequest(Request $request)
+    {
+        return array_replace_recursive($request->all(), $request->allFiles());
+    }
+
+    /**
+     * Get the file field name
+     *
+     * @return string|null
+     */
+    public static function getField(Request $request)
+    {
+        return array_key_first(Arr::dot(self::convertRequest($request)));
+    }
+
+    /**
      * Get the file from request
      *
      * @return \Illuminate\Http\UploadedFile|\Illuminate\Http\UploadedFile[]|null
      */
     public static function getUploadedFile(Request $request)
     {
-        $field = array_key_first(Arr::dot($request->all()));
+        $field = self::getField($request);
 
         return $request->file($field);
     }
 
     /**
-     * Get the file from request
+     * Get the metadata from request
      *
      * @return array|null
      */
     public static function getMetadata(Request $request)
     {
-        $field = array_key_first(Arr::dot($request->all()));
+        $field = self::getField($request);
 
         return json_decode($request->post($field) ?: '', true);
     }